add -v 4 for testing
Former-commit-id: 8eb27b9abfe4b38d51a3e65d81e2041ba3b3ec56

+++ virtual_machine/src/vm_params/vmp_player/file/file_code_size.c
@@ -10,13 +10,13 @@
 size_t      file_code_size(tt_file * self)
 {
     char		*code_size_in_data;
-    size_t		code_size;
+    unsigned int		code_size;
 
     code_size_in_data = (char*)(self->data)
         + COREWAR_EXEC_MAGIC_LENGTH
         + PROG_NAME_LENGTH
         + NULL_LENGTH;
-    code_size = (size_t)(*(int*)code_size_in_data);
+    code_size = (*(int*)code_size_in_data);
     code_size = (int)((code_size << 24)
         | ((code_size & 0xff00) << 8)
         | ((code_size & 0xff0000) >> 8)
